refactor: viewHolder 패턴 적용
